feat/Validating if task exist on database by id on all routes

+++ src/routes.js
@@ -48,6 +48,7 @@ export const routes = [
     path: buildRoutePath('/tasks/:id/complete'),
     handler: (req, res) => {
       const { id } = req.params
+
       const tasks = database.select('tasks')
       const task = tasks.find(task => task.id === id)
 
@@ -83,8 +84,8 @@ export const routes = [
 
       database.update('tasks', id, {
         ...task,
-        title,
-        description,
+        title: title ?? task.title, //Retorna task.title caso title seja nulo
+        description: description ?? task.description,
         updated_at: formattedDate()
       })
 
@@ -96,6 +97,13 @@ export const routes = [
     path: buildRoutePath('/tasks/:id'),
     handler: (req, res) => {
       const { id } = req.params
+
+      const tasks = database.select('tasks')
+      const task = tasks.find(task => task.id === id)
+
+      if (!task) {
+        return res.writeHead(404).end('Task not found.')
+      }
 
       database.delete('tasks', id)
